backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
pillow==12.0.0
PyJWT==2.10.1
redis==7.0.1
sqlparse==0.5.3
tzdata==2025.2
pytz==2025.2
PyYAML==6.0.3
uritemplate==4.2.0